+++ lib/git/pkgs.rb
@@ -33,6 +33,7 @@ require_relative "pkgs/commands/show"
 require_relative "pkgs/commands/log"
 require_relative "pkgs/commands/upgrade"
 require_relative "pkgs/commands/schema"
+require_relative "pkgs/commands/diff_driver"
 
 module Git
   module Pkgs